suggestion_service/requirements.txt
fastapi==0.115.1
uvicorn[standard]==0.31.1
pydantic==2.13.1
requests==2.32.3
redis==5.2.1
python-dotenv==1.0.1
